pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdproxy"
dynamic = ["version"]
authors = [
    { name = "limo", email = "limo.iie4@gmail.com" }
]
description = "A package for remote command execution"
readme = "README.md"
requires-python = ">=3.6"
keywords = ["dev", "utils", "remote"]
license = { file = "LICENSE" }
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Topic :: Software Development :: Build Tools',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = [
    "autoserde @ git+https://github.com/limoiie/autoserde.git@v0.0.1",
    "celery ~= 5.2.3",
    "fire ~= 0.4.0",
    "flexio @ git+https://github.com/limoiie/flexio.git@v0.0.1",
    "pymongo ~= 4.0.2",
    "pyyaml ~= 6.0",
    "redis ~= 3.5.3",
]

[project.optional-dependencies]
tests = ["pytest", "pytest-mock", "faker"]

[tool.setuptools_scm]